Move base extensible service
