+++ src/clims/services/base_extensible_service.py
@@ -9,6 +9,9 @@ from clims.services.extensible import ExtensibleTypeNotRegistered
 
 
 class BaseExtensibleService(object):
+    """
+    A base class for service classes that handle extensible objects.
+    """
 
     DoesNotExist = DoesNotExist
 
@@ -20,6 +23,9 @@ class BaseExtensibleService(object):
         self._archetype_version_class = wrapper_class.WrappedVersion
 
     def to_wrapper(self, model):
+        """
+        Wraps the ORM model to a higher level extensible class.
+        """
         if model is None:
             return None
         if isinstance(model, self._archetype_version_class):
@@ -30,6 +36,9 @@ class BaseExtensibleService(object):
             raise AssertionError("The model {} can't be wrapped".format(type(model)))
 
     def _version_to_wrapper(self, version):
+        """
+        Wraps the ORM model for a particular version to a higher level extensible class.
+        """
         try:
             SpecificExtensibleType = self._app.extensibles.get_implementation(
                 version.archetype.extensible_type.name)
@@ -146,9 +155,9 @@ class BaseExtensibleService(object):
             extensible_property_type__name=property).select_related('extensible_property_type')
         return property_query_set
 
-    @staticmethod
-    def get_values_of_property(property, extensible_type):
-        property_query_set = BaseExtensibleService._get_values_qs(property, extensible_type)
+    @classmethod
+    def get_values_of_property(cls, property, extensible_type):
+        property_query_set = cls._get_values_qs(property, extensible_type)
         try:
             # Note that this assumes that the property type is the same for all
             # entities in the property_query_set
@@ -160,10 +169,10 @@ class BaseExtensibleService(object):
             raise DoesNotExist("No property with name: {} found on extensible type: {}".format(property,
                                                                                                extensible_type))
 
-    @staticmethod
-    def get_unique_values_of_property(property, extensible_type):
+    @classmethod
+    def get_unique_values_of_property(cls, property, extensible_type):
         # TODO: add organization to the filter parameters
-        property_query_set = BaseExtensibleService._get_values_qs(property, extensible_type)
+        property_query_set = cls._get_values_qs(property, extensible_type)
         try:
             # Note that this assumes that the property type is the same for all
             # entities in the property_query_set
